DailyGoalActivity and DailyGoal classes are now implemented
