+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         <activity android:name=".ConsumptionActivity"></activity>
 
         <activity
-            android:name=".DailyGoalSettings"
+            android:name=".DailyGoalActivity"
             android:parentActivityName=".MainActivity"/>
         <activity android:name=".ShowList" />
         <activity android:name=".UnitActivity" />
@@ -23,6 +23,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <meta-data
+            android:name="preloaded_fonts"
+            android:resource="@array/preloaded_fonts" />
     </application>
 
 </manifest>
