+++ app/src/main/java/fi/metropolia/simppa/watertracker/DailyGoal.java
@@ -1,0 +1,18 @@
+package fi.metropolia.simppa.watertracker;
+
+public class DailyGoal {
+
+    private int dailygoal = 2500;
+
+    public DailyGoal(int dailygoal) {
+        this.dailygoal = dailygoal;
+    }
+
+    public void setDailygoal(int message) {
+        this.dailygoal = message;
+    }
+
+    public int getDailygoal() {
+        return this.dailygoal;
+    }
+}
