--- app/src/main/java/fi/metropolia/simppa/watertracker/DailyGoalSettings.java
@@ -1,38 +1,0 @@
-package fi.metropolia.simppa.watertracker;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-
-public class DailyGoalSettings extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        setContentView(R.layout.activity_goal);
-        Intent intent = getIntent();
-
-        hideNavigationBar();
-
-    }
-
-    @Override
-    protected void onResume(){
-        super.onResume();
-        hideNavigationBar();
-    }
-
-    private void hideNavigationBar() {
-        this.getWindow().getDecorView()
-                .setSystemUiVisibility(
-        View.SYSTEM_UI_FLAG_FULLSCREEN |
-                View.SYSTEM_UI_FLAG_HIDE_NAVIGATION |
-                View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY |
-                View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN |
-                View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION |
-                View.SYSTEM_UI_FLAG_LAYOUT_STABLE);
-    }
-}
