+++ app/src/main/java/fi/metropolia/simppa/watertracker/MainActivity.java
@@ -6,7 +6,6 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-import android.widget.ListView;
 
 public class MainActivity extends AppCompatActivity {
     Button b1, b2, b3;
@@ -30,7 +29,7 @@ public class MainActivity extends AppCompatActivity {
             intent = new Intent(this, ShowList.class);
             startActivity(intent);
         } else if (view.getId() == b3.getId()) {
-            intent = new Intent(this, DailyGoalSettings.class);
+            intent = new Intent(this, DailyGoalActivity.class);
             startActivity(intent);
         }
     }
