+++ app/src/main/res/layout/activity_goal.xml
@@ -5,16 +5,18 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".DailyGoalSettings">
+    tools:context=".DailyGoalActivity">
 
     <TextView
         android:id="@+id/textView2"
+
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="56dp"
+        android:fontFamily="@font/pontano_sans"
         android:text="@string/daily_goal"
         android:textColor="@color/icon"
-        android:textSize="40sp"
+        android:textSize="50sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
@@ -23,22 +25,24 @@
         android:id="@+id/textView3"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:fontFamily="@font/pontano_sans"
         android:layout_marginTop="32dp"
         android:text="@string/current_goal"
         android:textColor="@color/icon"
-        android:textSize="24sp"
+        android:textSize="30sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView2" />
 
     <TextView
-        android:id="@+id/textView4"
+        android:id="@+id/currentGoal"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:fontFamily="@font/pontano_sans"
         android:layout_marginTop="16dp"
         android:text="2500 ml"
         android:textColor="@color/icon"
-        android:textSize="24sp"
+        android:textSize="30sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView3" />
@@ -47,25 +51,28 @@
         android:id="@+id/textView5"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:fontFamily="@font/pontano_sans"
         android:layout_marginTop="32dp"
         android:text="@string/new_goal"
         android:textColor="@color/icon"
-        android:textSize="24sp"
+        android:textSize="30sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView4" />
+        app:layout_constraintTop_toBottomOf="@+id/currentGoal" />
 
     <EditText
         android:id="@+id/editText"
         android:layout_width="0dp"
-        android:layout_height="50dp"
+        android:layout_height="60dp"
         android:layout_marginStart="32dp"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="16dp"
         android:layout_marginEnd="32dp"
         android:background="@color/backgroundLite"
         android:ems="10"
+        android:fontFamily="@font/pontano_sans"
+        android:gravity="center_horizontal|center_vertical"
+        android:hint="2500"
         android:inputType="textPersonName"
-        android:text="2500"
         android:textColor="@android:color/darker_gray"
         android:textSize="20sp"
         app:layout_constraintEnd_toEndOf="parent"
@@ -73,16 +80,16 @@
         app:layout_constraintTop_toBottomOf="@+id/textView5" />
 
     <Button
-        android:id="@+id/button2"
+        android:id="@+id/buttonSetGoal"
         android:layout_width="0dp"
         android:layout_height="70dp"
         android:layout_marginStart="76dp"
         android:layout_marginEnd="76dp"
-
         android:background="@drawable/custom_button"
-    android:textAllCaps="false"
+        android:fontFamily="@font/pontano_sans"
         android:text="@string/set_goal"
-        android:textSize="20sp"
+        android:textAllCaps="false"
+        android:textSize="24sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
